requirements.txt
pyrogram==2.0.106
pytgcalls==3.0.0.dev6
yt-dlp==2023.11.16
PyYAML==6.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
pymongo==4.6.1
motor==3.3.2
